fix: preserve daemon status contract anchors
